Change the color of the sinkIndicator while LinkItem is pending
I think displaying the actual position of the dataflow is very useful, but the yellow color was so hard to notice that many people didn't even know about this feature. Red is definitely visible (although not colorblind friendly), but now it looks a bit funny because the sinkIndicator layer is below the shield. It could be considered to change the color of the whole input shield while the flow is pending or move the sinkIndicator layer above the shield.

Transplanted-from: biolab/orange3@aab008b994acff0e3c601866d7aac485e6b4a276