quickmenu: Fix a segfault in up/down keyboard navigation ...
... when the mouse cursor is above the view and the current index is
moved above or below the viewport.

This is due to the use of 'entered' signal from the view which gets
emitted when the view scrolls the view to the current index when it
is moved outside the viewport bounds (via up/down keys).

The `__onEntered` slot then sets the current index again,
before the view fully processed the previous move (the `__onEntered`
slot already contains one note and a 'workaround' for a similar or
maybe even exact issue).

Replace the use of 'entered' signal from the view with a mouse move
event filter on the view.